requirements.txt
astroid==2.2.3
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==3.2.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
pylint==2.3.1
python-dotenv==0.10.1
python-memcached==1.59
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.3.1
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
